Update core installations and rename some files
New file coreDependencies to isntall new and updated packages versions and all what electron needs, also was renamed some python files to comply with best practices of python
